requirements.txt
# langchain==0.0.195 # Doesn't work!!
langchain==0.0.191
# langchain==0.0.177
gpt4all==0.3.2
# gpt4all==0.2.3
chromadb==0.3.26
# chromadb==0.3.23
llama-cpp-python==0.1.61
# llama-cpp-python==0.1.50
urllib3==2.0.3
# urllib3==2.0.2
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.7.3
# unstructured==0.6.6
extract-msg==0.41.2
# extract-msg==0.41.1
tabulate==0.9.0
pandas==2.0.2
pandoc==2.3
pypandoc==1.11
tqdm==4.65.0
docx2txt==0.8
sqlalchemy==2.0.15
# nltk==9.999
safetensors==0.3.1
# apex helps optimise inference on T5 models
#apex==  # counldn't install the cryptacular package!!
# Required for Huggingface training.
xformers==0.0.20
transformers==4.30.1
tensorflow==2.12.0
